% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{camellia_flower}
\alias{camellia_flower}
\title{동백꽃}
\format{UTF-8 문자열 텍스트.}
\source{
인터넷 검색
}
\usage{
camellia_flower
}
\description{
김유정의 단편소설. 시골 남녀의 풋풋한 사랑을 그려낸 작품으로, 1936년에 발표되었다.
}
\keyword{datasets}
