% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three_generation}
\alias{three_generation}
\title{삼대}
\format{UTF-8 문자열 텍스트.}
\source{
인터넷 검색
}
\usage{
three_generation
}
\description{
소설가 염상섭의 장편 연작소설.
총 42장으로 구성되어 있으며 만세전과 함께 염상섭 문학을 대표하는 작품이다.
}
\keyword{datasets}
